@@ -1,23 +1,24 @@
+import io.gitlab.arturbosch.detekt.Detekt
+
 apply plugin: 'com.android.application'
 apply plugin: 'kotlin-android'
-apply plugin: 'kotlin-android-extensions'
 apply plugin: 'kotlin-kapt'
 apply plugin: "io.gitlab.arturbosch.detekt"
-
+apply plugin: 'kotlin-parcelize'
 
 android {
 
-    compileSdkVersion 29
+    compileSdk 34
+
     defaultConfig {
         applicationId "education.mahmoud.quranyapp"
         minSdkVersion 21
-        targetSdkVersion 29
-        versionCode 30
-        versionName "3.0.0"
+        targetSdkVersion 34
+        versionCode 31
+        versionName "3.0.1"
         multiDexEnabled true
-//        testInstrumentationRumannner "android.support.test.runner.AndroidJUnitRunner"
-        vectorDrawables.useSupportLibrary = true
     }
+
     buildTypes {
         release {
             minifyEnabled true
@@ -27,45 +28,46 @@ android {
     }
     sourceSets { main { assets.srcDirs = ['src/main/assets', 'src/main/assets/'] } }
     compileOptions {
-        sourceCompatibility = '1.8'
-        targetCompatibility = '1.8'
+        sourceCompatibility = JavaVersion.VERSION_17
+        targetCompatibility = JavaVersion.VERSION_17
+    }
+    kotlinOptions {
+        jvmTarget = JavaVersion.VERSION_17
     }
     testOptions {
         unitTests.returnDefaultValues = true
     }
-    dataBinding {
-        enabled = true
+
+    buildFeatures{
+        viewBinding = true
+        dataBinding = true
     }
+    namespace 'education.mahmoud.quranyapp'
 }
 
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
 
-    testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'androidx.test:runner:1.4.0-beta02'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0-beta02'
+    testImplementation 'junit:junit:4.13.2'
+    androidTestImplementation 'androidx.test:runner:1.5.2'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
 
-
-    // kotlin
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
-
-
-    def support_version = '1.3.0'
+    def support_version = '1.6.1'
     implementation "androidx.appcompat:appcompat:$support_version"
-    implementation "com.google.android.material:material:$support_version"
+    def material_version = '1.12.0'
+    implementation "com.google.android.material:material:$material_version"
 
 
-    implementation 'androidx.core:core-ktx:1.5.0'
-    implementation 'androidx.constraintlayout:constraintlayout:2.0.4'
+    implementation 'androidx.core:core-ktx:1.13.1'
+    implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
     implementation 'androidx.legacy:legacy-support-v4:1.0.0'
 
     implementation 'com.android.support:multidex:1.0.3'
 
     implementation 'com.jakewharton:butterknife:10.1.0'
     annotationProcessor "com.jakewharton:butterknife-compiler:10.1.0"
-    //   implementation "com.balsikandar.android:crashreporter:1.0.9"
 
-    implementation 'com.google.code.gson:gson:2.8.5'
+    implementation 'com.google.code.gson:gson:2.10.1'
 
     implementation 'com.intuit.sdp:sdp-android:1.0.6'
     implementation 'com.intuit.ssp:ssp-android:1.0.6'
