@@ -1,19 +1,21 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
+
+
 buildscript {
 
     ext{
-        kotlin_version = '1.6.0'
+        kotlin_version = '1.8.22'
         rxjava_version = '1.3.0'
         rxandroid_version = '1.2.1'
         multidex_version = '1.0.3'
         rxfirebase_version = '1.5.0'
-        lifecycle_version = '2.2.0'
+        lifecycle_version = '2.7.0'
     }
 
     repositories {
         google()
-        jcenter()
+        mavenCentral()
         maven { url 'https://jitpack.io' }
         maven {
             url "https://plugins.gradle.org/m2/"
@@ -21,7 +23,7 @@ buildscript {
 
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:4.2.2'
+        classpath 'com.android.tools.build:gradle:8.2.2'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
         classpath "org.jlleitschuh.gradle:ktlint-gradle:7.1.0"
         
@@ -31,6 +33,14 @@ buildscript {
 
         classpath "io.gitlab.arturbosch.detekt:detekt-gradle-plugin:1.18.1"
     }
+}
+
+//plugins {
+//    id 'com.google.devtools.ksp' version '1.8.10-1.0.9' apply false
+//}
+
+plugins {
+    id "org.jetbrains.kotlin.android" version "1.4.20" apply false
 }
 
 allprojects {
@@ -45,6 +55,7 @@ allprojects {
     }
 }
 
+
 task clean(type: Delete) {
     delete rootProject.buildDir
 }
@@ -52,4 +63,6 @@ task clean(type: Delete) {
 repositories {
     mavenCentral()
 }
+
+
 
