@@ -8,12 +8,11 @@ import android.widget.AdapterView.OnItemSelectedListener
 import android.widget.ArrayAdapter
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
-import butterknife.ButterKnife
 import education.mahmoud.quranyapp.R
+import education.mahmoud.quranyapp.databinding.ActivitySettingBinding
 import education.mahmoud.quranyapp.datalayer.QuranRepository
-import kotlinx.android.synthetic.main.activity_setting.*
+import education.mahmoud.quranyapp.utils.viewBinding
 import org.koin.android.ext.android.inject
-import java.util.*
 
 class SettingActivity : AppCompatActivity() {
 
@@ -21,57 +20,76 @@ class SettingActivity : AppCompatActivity() {
 
     val quranRepository: QuranRepository by inject()
 
+    private val binding by viewBinding(ActivitySettingBinding::inflate)
+
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_setting)
-        ButterKnife.bind(this)
-        val colorSet: List<String> = ArrayList(listOf(getString(R.string.white), getString(R.string.yellow), getString(R.string.green)))
-        cbNightMode.isChecked = quranRepository.nightModeState
+        setContentView(binding.root)
+        with(binding) {
+            val colorSet: List<String> = ArrayList(
+                listOf(
+                    getString(R.string.white),
+                    getString(R.string.yellow),
+                    getString(R.string.green)
+                )
+            )
+            cbNightMode.isChecked = quranRepository.nightModeState
 
-        if (cbNightMode.isChecked) {
-            nightMode()
-        } else {
-            defaultMode()
-        }
-
-        cbNightMode.setOnCheckedChangeListener { compoundButton, state ->
-            Log.d(TAG, "onCheckedChanged: $state")
-            quranRepository.nightModeState = state
-            if (state) {
+            if (cbNightMode.isChecked) {
                 nightMode()
             } else {
                 defaultMode()
             }
-        }
-        val adapter = ArrayAdapter(this, android.R.layout.simple_dropdown_item_1line, colorSet)
-        spColorReqularMode.adapter = adapter
-        // load from shared preference and set to spinner.
-        colorPosForBackground = quranRepository.backColorState
-        spColorReqularMode.setSelection(colorPosForBackground)
-        spColorReqularMode.onItemSelectedListener = object : OnItemSelectedListener {
-            override fun onItemSelected(adapterView: AdapterView<*>?, view: View, i: Int, l: Long) {
-                Log.d(TAG, "onItemSelected: $i:: $l")
-                val pos = spColorReqularMode.selectedItemPosition
-                Log.d(TAG, "onItemSelected: pos $pos")
-                colorPosForBackground = pos
+
+            cbNightMode.setOnCheckedChangeListener { compoundButton, state ->
+                Log.d(TAG, "onCheckedChanged: $state")
+                quranRepository.nightModeState = state
+                if (state) {
+                    nightMode()
+                } else {
+                    defaultMode()
+                }
             }
+            val adapter = ArrayAdapter(
+                this@SettingActivity,
+                android.R.layout.simple_dropdown_item_1line,
+                colorSet
+            )
+            spColorReqularMode.adapter = adapter
+            // load from shared preference and set to spinner.
+            colorPosForBackground = quranRepository.backColorState
+            spColorReqularMode.setSelection(colorPosForBackground)
+            spColorReqularMode.onItemSelectedListener = object : OnItemSelectedListener {
+                override fun onItemSelected(
+                    adapterView: AdapterView<*>?,
+                    view: View,
+                    i: Int,
+                    l: Long,
+                ) {
+                    Log.d(TAG, "onItemSelected: $i:: $l")
+                    val pos = spColorReqularMode.selectedItemPosition
+                    Log.d(TAG, "onItemSelected: pos $pos")
+                    colorPosForBackground = pos
+                }
 
-            override fun onNothingSelected(adapterView: AdapterView<*>?) {}
-        }
-        title = getString(R.string.setting)
+                override fun onNothingSelected(adapterView: AdapterView<*>?) {}
+            }
+            title = getString(R.string.setting)
 
-        btnSetColor.setOnClickListener {
-            quranRepository.backColorState = colorPosForBackground
-            showMessage(getString(R.string.setting_updated))
+            btnSetColor.setOnClickListener {
+                quranRepository.backColorState = colorPosForBackground
+                showMessage(getString(R.string.setting_updated))
+            }
         }
     }
 
     private fun defaultMode() {
-        linearColor.visibility = View.VISIBLE
+        binding.linearColor.visibility = View.VISIBLE
     }
 
     private fun nightMode() {
-        linearColor.visibility = View.GONE
+        binding.linearColor.visibility = View.GONE
     }
 
     private fun showMessage(message: String) {
