@@ -8,19 +8,25 @@ import android.view.ViewGroup
 import android.widget.AdapterView
 import android.widget.AdapterView.OnItemSelectedListener
 import android.widget.ArrayAdapter
+import com.zhuinden.fragmentviewbindingdelegatekt.viewBinding
 import education.mahmoud.quranyapp.R
 import education.mahmoud.quranyapp.base.BaseFragment
+import education.mahmoud.quranyapp.databinding.FragmentTafseerDetailsBinding
 import education.mahmoud.quranyapp.datalayer.QuranRepository
-import kotlinx.android.synthetic.main.fragment_tafseer_details.*
 import org.koin.android.ext.android.inject
-import java.util.*
 
 class TafseerDetails : BaseFragment() {
     private val quranRepository: QuranRepository by inject()
 
     private val adapter = TafseerAdapter()
 
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
+    private val binding by viewBinding(FragmentTafseerDetailsBinding::bind)
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?,
+    ): View? {
         return inflater.inflate(R.layout.fragment_tafseer_details, container, false)
     }
 
@@ -32,47 +38,66 @@ class TafseerDetails : BaseFragment() {
 
     override fun setClickListeners() {
         super.setClickListeners()
-        tvNoDataInTafseer.setOnClickListener {}
+        binding.tvNoDataInTafseer.setOnClickListener {}
     }
 
     private fun initRv() {
-        rvTafeer.setHasFixedSize(true)
-        rvTafeer.adapter = (adapter)
+        with(binding) {
+            rvTafeer.setHasFixedSize(true)
+            rvTafeer.adapter = (adapter)
+        }
     }
 
     private fun fillSpinners() {
         val suraNames = quranRepository.surasNames
-        val startAdapter = ArrayAdapter(requireContext(), android.R.layout.simple_dropdown_item_1line, suraNames)
-        spSuraTafser.adapter = startAdapter
-        try {
-            spSuraTafser.onItemSelectedListener = object : OnItemSelectedListener {
-                override fun onItemSelected(adapterView: AdapterView<*>?, view: View?, i: Int, l: Long) {
-                    val ayahs = quranRepository.getAllAyahOfSurahIndexForTafseer(l + 1)
-                    Log.d(TAG, "onItemSelected: " + ayahs.size + " & " + l)
-                    // update adapter
-                    if (ayahs.isNotEmpty()) {
-                        foundState()
-                        adapter.setTafseerList(ayahs)
-                        // Log.d(TAG, "onItemSelected: " + suraName);
-                        val ayahsNums = creatAyahsNumList(ayahs.size) // size() ->  n of ayahs
-                        val adapter = ArrayAdapter(requireContext(), android.R.layout.simple_dropdown_item_1line, ayahsNums)
-                        spAyahTafser.adapter = adapter
-                        spAyahTafser.onItemSelectedListener = object : OnItemSelectedListener {
-                            override fun onItemSelected(adapterView: AdapterView<*>?, view: View?, i: Int, l: Long) {
-                                rvTafeer.scrollToPosition(l.toInt())
-                                Log.d(TAG, "onItemSelected: TAFSEEER POS $l")
+        val startAdapter =
+            ArrayAdapter(requireContext(), android.R.layout.simple_dropdown_item_1line, suraNames)
+        with(binding) {
+            spSuraTafser.adapter = startAdapter
+            try {
+                spSuraTafser.onItemSelectedListener = object : OnItemSelectedListener {
+                    override fun onItemSelected(
+                        adapterView: AdapterView<*>?,
+                        view: View?,
+                        i: Int,
+                        l: Long,
+                    ) {
+                        val ayahs = quranRepository.getAllAyahOfSurahIndexForTafseer(l + 1)
+                        Log.d(TAG, "onItemSelected: " + ayahs.size + " & " + l)
+                        // update adapter
+                        if (ayahs.isNotEmpty()) {
+                            foundState()
+                            adapter.setTafseerList(ayahs)
+                            // Log.d(TAG, "onItemSelected: " + suraName);
+                            val ayahsNums = creatAyahsNumList(ayahs.size) // size() ->  n of ayahs
+                            val adapter = ArrayAdapter(
+                                requireContext(),
+                                android.R.layout.simple_dropdown_item_1line,
+                                ayahsNums
+                            )
+                            spAyahTafser.adapter = adapter
+                            spAyahTafser.onItemSelectedListener = object : OnItemSelectedListener {
+                                override fun onItemSelected(
+                                    adapterView: AdapterView<*>?,
+                                    view: View?,
+                                    i: Int,
+                                    l: Long,
+                                ) {
+                                    rvTafeer.scrollToPosition(l.toInt())
+                                    Log.d(TAG, "onItemSelected: TAFSEEER POS $l")
+                                }
+
+                                override fun onNothingSelected(adapterView: AdapterView<*>?) {}
                             }
-
-                            override fun onNothingSelected(adapterView: AdapterView<*>?) {}
+                        } else {
+                            notFoundState()
                         }
-                    } else {
-                        notFoundState()
                     }
-                }
 
-                override fun onNothingSelected(adapterView: AdapterView<*>?) {}
+                    override fun onNothingSelected(adapterView: AdapterView<*>?) {}
+                }
+            } catch (e: Exception) {
             }
-        } catch (e: Exception) {
         }
     }
 
@@ -85,17 +110,21 @@ class TafseerDetails : BaseFragment() {
     }
 
     private fun foundState() {
-        rvTafeer.visibility = View.VISIBLE
-        spSuraTafser.visibility = View.VISIBLE
-        spAyahTafser.visibility = View.VISIBLE
-        tvNoDataInTafseer.visibility = View.GONE
+        with(binding) {
+            rvTafeer.visibility = View.VISIBLE
+            spSuraTafser.visibility = View.VISIBLE
+            spAyahTafser.visibility = View.VISIBLE
+            tvNoDataInTafseer.visibility = View.GONE
+        }
     }
 
     private fun notFoundState() {
-        rvTafeer.visibility = View.GONE
-        spSuraTafser.visibility = View.INVISIBLE
-        spAyahTafser.visibility = View.INVISIBLE
-        tvNoDataInTafseer.visibility = View.VISIBLE
+        with(binding) {
+            rvTafeer.visibility = View.GONE
+            spSuraTafser.visibility = View.INVISIBLE
+            spAyahTafser.visibility = View.INVISIBLE
+            tvNoDataInTafseer.visibility = View.VISIBLE
+        }
     }
 
     companion object {
